// apps/frontend/lib/types.ts
export type SignUpResponse =
  | { status: "success" }
  | {
      status: "validation_error";
      errors: [
        InferIssue<typeof SignUpSchema>,
        ...InferIssue<typeof SignUpSchema>[],
      ];
    }
  | { status: "server_error"; error: string };